Move connection_utils to app/test_utils
